ci: publish Docker Image on release
Move the docker build and push logic to a reusable workflow,
and use in both build and release-please (if a release were created).

Since the tag is created in a workflow, it does not trigger the build
workflow. The release thus needs to be done in release-please workflow.
This means the version must be passed to the docker-metadata action
(since the ref won't be tag/*).

Also, ensure the build workflow is not triggered on new release
(using paths-ignore), and skip the docker publication if the
commit is 'ci:'.