feat!: add optional URL shortening (requires persistence)
To make it less obvious the link is "fishy", disguise rickroller
into an URL shortener. This requires a database.

Support SQL-like backend (sqlachemy), but also no persistence
and SQLLite in-memory for testing.

To avoid the db to grow too much, a job is registered to run
every so often and clear entries that have not been accessed
in a while.