fix: support log level DEBUG
To be able to control the logging level from the loggers,
it is important the logging level of the handlers is either
set to NOTSET or to a level similar or lower.
For example, if a handler has 'warning', setting 'debug' on the
logger using it has no effect.

Hence, remove all logging level information from the handlers.

Also: put back the "main()" method, so we can run rickroll
using the syntax `poetry run rickroll`.