fix: use a custom page for rickroll instead of YouTube
YouTube redirect is not ideal: opens an app on mobile, may start
with an ad, autoplay not always working, etc.

So instead, host a "rolled" page which uses an embed from giphy.
The sound doesn't play, but this is to make the autoplay feature
work (autoplay with sounds is restricted since Chrome 66).